+++ packages/sdk-core/src/ERC20Token.ts
@@ -46,7 +46,7 @@ export default class ERC20Token {
             return allowance.toString();
         } catch (err) {
             throw new SFError({
-                code: "SUPERTOKEN_READ",
+                type: "SUPERTOKEN_READ",
                 message: "There was an error getting allowance",
                 cause: err,
             });
@@ -74,7 +74,7 @@ export default class ERC20Token {
             return balanceOf.toString();
         } catch (err) {
             throw new SFError({
-                code: "SUPERTOKEN_READ",
+                type: "SUPERTOKEN_READ",
                 message: "There was an error getting balanceOf",
                 cause: err,
             });
@@ -96,7 +96,7 @@ export default class ERC20Token {
             return name;
         } catch (err) {
             throw new SFError({
-                code: "SUPERTOKEN_READ",
+                type: "SUPERTOKEN_READ",
                 message: "There was an error getting name",
                 cause: err,
             });
@@ -120,7 +120,7 @@ export default class ERC20Token {
             return symbol;
         } catch (err) {
             throw new SFError({
-                code: "SUPERTOKEN_READ",
+                type: "SUPERTOKEN_READ",
                 message: "There was an error getting symbol",
                 cause: err,
             });
@@ -144,7 +144,7 @@ export default class ERC20Token {
             return totalSupply.toString();
         } catch (err) {
             throw new SFError({
-                code: "SUPERTOKEN_READ",
+                type: "SUPERTOKEN_READ",
                 message: "There was an error getting totalSupply",
                 cause: err,
             });