+++ packages/sdk-core/CHANGELOG.md
@@ -10,8 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Remove serialized internal error from the messages of `SFError` (it's now just included in the `.cause` property)
 
 ### Breaking
-- `ErrorType` renamed to `ErrorCode`
-- `SFError.type` renamed to `SFError.code`  
+- `SFError.errorObject` renamed to `SFError.cause`  
 
 ## [0.4.4] - 2022-06-30